# Remove overhead of surrounding block from length-vs-size-vs-count benchmark