@nateberkopec 2016-01-25
:+1: We should probably write something down, maybe in Contributing.md, about preferring this style where possible.